… pkg/genlib/generator_test.go
… pkg/genlib/generator_with_custom_template_test.go
… pkg/genlib/generator_with_text_template.go
… pkg/genlib/generator_with_text_template_test.go