`main` branch:
```
goos: darwin
goarch: amd64
pkg: github.com/elastic/elastic-integration-corpus-generator-tool/pkg/genlib
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_GeneratorCustomTemplateJSONContent-16    	  235588	     47286 ns/op	     432 B/op	      14 allocs/op
Benchmark_GeneratorTextTemplateJSONContent-16      	   24507	    487438 ns/op	   48725 B/op	    2252 allocs/op
Benchmark_GeneratorCustomTemplateVPCFlowLogs-16    	 6558933	      1840 ns/op	      64 B/op	       2 allocs/op
Benchmark_GeneratorTextTemplateVPCFlowLogs-16      	  526669	     22920 ns/op	    2322 B/op	      95 allocs/op
PASS
ok  	github.com/elastic/elastic-integration-corpus-generator-tool/pkg/genlib	57.004s
```


this branch:
```
goos: darwin
goarch: amd64
pkg: github.com/elastic/elastic-integration-corpus-generator-tool/pkg/genlib
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
Benchmark_GeneratorCustomTemplateJSONContent-16    	   97422	    145236 ns/op	   18980 B/op	     972 allocs/op
Benchmark_GeneratorTextTemplateJSONContent-16      	   18703	    820674 ns/op	   49159 B/op	    2265 allocs/op
Benchmark_GeneratorCustomTemplateVPCFlowLogs-16    	 1504225	      8486 ns/op	     501 B/op	      35 allocs/op
Benchmark_GeneratorTextTemplateVPCFlowLogs-16      	  299754	     43765 ns/op	    2540 B/op	     111 allocs/op
PASS
ok  	github.com/elastic/elastic-integration-corpus-generator-tool/pkg/genlib	73.641s
```