@aspacca 2023-02-01
@endorama you were right

[tweaking the channel size](https://github.com/elastic/elastic-integration-corpus-generator-tool/pull/51/files#diff-0fc87380e8715fdb1a33c600e6513325ba82801e849851c5856064b7beac59b8R35-R40) seems to show different performance behaviour and some improvement, but I'd say it's not reliable, since it's probably very relevant to the host machine